requirements.txt
scipy==1.5.2
numpy==1.19.2
tensorboard
pymilvus==0.2.14
torch==1.6.0 # install with conda instead https://pytorch.org/
tqdm==4.49.0
torch_optimizer==0.0.1a14
matplotlib==3.3.1
jupyterthemes==0.20.0
pandas==1.1.2
scikit_learn==0.23.2
sphinx_rtd_theme==0.5.0
streamlit==0.66.0
pytest==6.0.2
